+++ packages/server-generator-apollo-emmett/CHANGELOG.md
@@ -1,11 +1,16 @@
 # @auto-engineer/server-generator-apollo-emmett
 
-## 0.1.5
+## 0.2.0
+
+### Minor Changes
+
+- add command details in dashboard
 
 ### Patch Changes
 
 - Updated dependencies
   - @auto-engineer/flow@0.2.0
+  - @auto-engineer/message-bus@0.6.0
 
 ## 0.1.4